+++ audionimbus/src/effect/direct.rs
@@ -126,57 +126,18 @@ pub struct DirectEffectParams {
 
 impl From<audionimbus_sys::IPLDirectEffectParams> for DirectEffectParams {
     fn from(params: audionimbus_sys::IPLDirectEffectParams) -> Self {
-        let distance_attenuation = if params.flags
-            & audionimbus_sys::IPLDirectEffectFlags::IPL_DIRECTEFFECTFLAGS_APPLYDISTANCEATTENUATION
-            == audionimbus_sys::IPLDirectEffectFlags::IPL_DIRECTEFFECTFLAGS_APPLYDISTANCEATTENUATION
-        {
-            Some(params.distanceAttenuation)
-        } else {
-            None
-        };
-
-        let air_absorption = if params.flags
-            & audionimbus_sys::IPLDirectEffectFlags::IPL_DIRECTEFFECTFLAGS_APPLYAIRABSORPTION
-            == audionimbus_sys::IPLDirectEffectFlags::IPL_DIRECTEFFECTFLAGS_APPLYAIRABSORPTION
-        {
-            Some(Equalizer(params.airAbsorption))
-        } else {
-            None
-        };
-
-        let directivity = if params.flags
-            & audionimbus_sys::IPLDirectEffectFlags::IPL_DIRECTEFFECTFLAGS_APPLYDIRECTIVITY
-            == audionimbus_sys::IPLDirectEffectFlags::IPL_DIRECTEFFECTFLAGS_APPLYDIRECTIVITY
-        {
-            Some(params.directivity)
-        } else {
-            None
-        };
-
-        let occlusion = if params.flags
-            & audionimbus_sys::IPLDirectEffectFlags::IPL_DIRECTEFFECTFLAGS_APPLYOCCLUSION
-            == audionimbus_sys::IPLDirectEffectFlags::IPL_DIRECTEFFECTFLAGS_APPLYOCCLUSION
-        {
-            Some(params.occlusion)
-        } else {
-            None
-        };
-
-        let transmission = if params.flags
-            & audionimbus_sys::IPLDirectEffectFlags::IPL_DIRECTEFFECTFLAGS_APPLYTRANSMISSION
-            == audionimbus_sys::IPLDirectEffectFlags::IPL_DIRECTEFFECTFLAGS_APPLYTRANSMISSION
-        {
-            Some(match params.transmissionType {
-                audionimbus_sys::IPLTransmissionType::IPL_TRANSMISSIONTYPE_FREQINDEPENDENT => {
-                    Transmission::FrequencyIndependent(Equalizer(params.transmission))
-                }
-                audionimbus_sys::IPLTransmissionType::IPL_TRANSMISSIONTYPE_FREQDEPENDENT => {
-                    Transmission::FrequencyDependent(Equalizer(params.transmission))
-                }
-            })
-        } else {
-            None
-        };
+        let distance_attenuation = Some(params.distanceAttenuation);
+        let air_absorption = Some(Equalizer(params.airAbsorption));
+        let directivity = Some(params.directivity);
+        let occlusion = Some(params.occlusion);
+        let transmission = Some(match params.transmissionType {
+            audionimbus_sys::IPLTransmissionType::IPL_TRANSMISSIONTYPE_FREQINDEPENDENT => {
+                Transmission::FrequencyIndependent(Equalizer(params.transmission))
+            }
+            audionimbus_sys::IPLTransmissionType::IPL_TRANSMISSIONTYPE_FREQDEPENDENT => {
+                Transmission::FrequencyDependent(Equalizer(params.transmission))
+            }
+        });
 
         Self {
             distance_attenuation,